[FIX] project_ux: Proper filter the subtasks (#81)
Formerly when copuy the project we was using a method to filter the subtasks
that was not taking into account the tasks in the original template project
that were in an stage folded or archived. The result of this was that we
were moving the task in folded stage from the template project to the new
project.

To avoid this we filter the subtasks as the one related to the project that
have parent_id.